pyproject.toml
[tool.poetry]
name = "bug-report"
version = "0.1.0"
description = ""
authors = ["Risabh Sharma <99526730+sharma-rishabh@users.noreply.github.com >"]
packages = [{include = "bug_report"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
requests = "^2.28.1"
python-dotenv = "^0.21.0"
celery = "^5.2.7"
redis = "^4.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
